refactor: extract hubspot sync logic into separate hook
